requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pymongo>=4.6.0
alembic>=1.13.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.6.0
google-generativeai>=0.3.0
pandas>=2.3.3
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.15.0
torch>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.0
openpyxl>=3.1.2
python-multipart>=0.0.6
pillow>=10.2.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

